Order update and create totalPrice feature hotfix;

+++ src/main/java/com/sergosoft/productservice/domain/order/Order.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Value;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.List;
 import java.util.UUID;
@@ -20,6 +21,6 @@ public class Order {
     UUID sellerId;
     UUID buyerId;
 
-    Double totalPrice;
+    BigDecimal totalPrice;
     Instant createdAt;
 }
